quiting voice_channel only if the queue is empty

+++ src/lib/src/application/infra/lavalink_ctx.rs
@@ -33,15 +33,29 @@ impl LavalinkHandler {
         client: LavalinkClient,
         guild_id: u64,
     ) -> Result<(), Error> {
-        self.songbird.remove(guild_id).await?;
+        let empty = {
+            let node = client.nodes().await;
 
-        let nodes = client.nodes().await;
-        nodes.remove(&guild_id);
+            let guild_node = node
+                .get(&guild_id)
+                .ok_or_else(|| anyhow!("Can't get guild loop for {}", guild_id))?;
 
-        let loops = client.loops().await;
-        loops.remove(&guild_id);
+            println!("Fila: {:?}", guild_node.queue);
 
-        client.destroy(guild_id).await?;
+            guild_node.queue.is_empty()
+        };
+
+        if empty {
+            self.songbird.remove(guild_id).await?;
+
+            let nodes = client.nodes().await;
+            nodes.remove(&guild_id);
+
+            let loops = client.loops().await;
+            loops.remove(&guild_id);
+
+            client.destroy(guild_id).await?;
+        }
 
         Ok(())
     }
